+++ modules/Attendance/AddAbsences.php
@@ -323,7 +323,7 @@ if ( ! $_REQUEST['modfunc'] )
 	$extra['functions'] = [ 'CHECKBOX' => 'MakeChooseCheckbox' ];
 
 	$extra['columns_before'] = [
-		'CHECKBOX' => MakeChooseCheckbox( '', 'STUDENT_ID', 'student' ),
+		'CHECKBOX' => MakeChooseCheckbox( 'required', 'STUDENT_ID', 'student' ),
 	];
 
 	$extra['new'] = true;