+++ modules/Custom/AttendanceSummary.php
@@ -316,7 +316,7 @@ if ( ! $_REQUEST['modfunc'] )
 
 	$extra['functions'] = [ 'CHECKBOX' => 'MakeChooseCheckbox' ];
 
-	$extra['columns_before'] = [ 'CHECKBOX' => MakeChooseCheckbox( 'Y', '', 'st_arr' ) ];
+	$extra['columns_before'] = [ 'CHECKBOX' => MakeChooseCheckbox( 'Y_required', '', 'st_arr' ) ];
 
 	$extra['options']['search'] = false;