+++ modules/Custom/RemoveAccess.php
@@ -161,7 +161,7 @@ if ( ! $_REQUEST['modfunc'] )
 	$extra['SELECT'] .= ",s.STUDENT_ID AS CHECKBOX,s.USERNAME";
 	$extra['link'] = [ 'FULL_NAME' => false ];
 	$extra['functions'] = [ 'CHECKBOX' => 'MakeChooseCheckbox' ];
-	$extra['columns_before'] = [ 'CHECKBOX' => MakeChooseCheckbox( 'Y', '', 'st_arr' ) ];
+	$extra['columns_before'] = [ 'CHECKBOX' => MakeChooseCheckbox( 'Y_required', '', 'st_arr' ) ];
 	$extra['new'] = true;
 
 	$extra['action'] = '&accessfunc=' . $accessfunc;