+++ modules/Eligibility/AddActivity.php
@@ -89,7 +89,7 @@ if ( $_REQUEST['search_modfunc'] === 'list' )
 $extra['link'] = [ 'FULL_NAME' => false ];
 $extra['SELECT'] = ",NULL AS CHECKBOX";
 $extra['functions'] = [ 'CHECKBOX' => 'MakeChooseCheckbox' ];
-$extra['columns_before'] = [ 'CHECKBOX' => MakeChooseCheckbox( '', 'STUDENT_ID', 'student' ) ];
+$extra['columns_before'] = [ 'CHECKBOX' => MakeChooseCheckbox( 'required', 'STUDENT_ID', 'student' ) ];
 $extra['new'] = true;
 Widgets( 'activity' );
 Widgets( 'course' );