+++ modules/Food_Service/Users/Reminders.php
@@ -109,7 +109,7 @@ if ( ! $_REQUEST['modfunc'] || $_REQUEST['search_modfunc'] === 'list' )
 	$extra['link'] = [ 'FULL_NAME' => false ];
 	$extra['SELECT'] = ",s.STAFF_ID AS CHECKBOX";
 	$extra['functions'] = [ 'CHECKBOX' => '_makeChooseCheckbox' ];
-	$extra['columns_before'] = [ 'CHECKBOX' => MakeChooseCheckbox( 'Y', '', 'st_arr' ) ];
+	$extra['columns_before'] = [ 'CHECKBOX' => MakeChooseCheckbox( 'Y_required', '', 'st_arr' ) ];
 	$extra['new'] = true;
 	$extra['options']['search'] = false;