+++ modules/Grades/ReportCards.php
@@ -90,7 +90,7 @@ if ( ! $_REQUEST['modfunc'] )
 	$extra['functions'] = [ 'CHECKBOX' => 'MakeChooseCheckbox' ];
 
 	$extra['columns_before'] = [
-		'CHECKBOX' => MakeChooseCheckbox( 'Y', '', 'st_arr' )
+		'CHECKBOX' => MakeChooseCheckbox( 'Y_required', '', 'st_arr' )
 	];
 
 	$extra['options']['search'] = false;