+++ modules/Students/AssignOtherInfo.php
@@ -453,7 +453,7 @@ if ( ! $_REQUEST['modfunc'] )
 	//Widgets('absences');
 
 	$extra['functions'] = [ 'CHECKBOX' => 'MakeChooseCheckbox' ];
-	$extra['columns_before'] = [ 'CHECKBOX' => MakeChooseCheckbox( '', 'STUDENT_ID', 'student' ) ];
+	$extra['columns_before'] = [ 'CHECKBOX' => MakeChooseCheckbox( 'required', 'STUDENT_ID', 'student' ) ];
 	$extra['new'] = true;
 
 	Search( 'student_id', $extra );